Parse Any in regex (`.`)
This was more complicated than I expected, because it collides with how variables are parsed.

Variables are now lazy, which means they will parse as few characters as possible, while a `.+` will parse as many characters as possible.

Otherwise regexes like `{foo}b{bar}` would be ambiguos, because regex could not know when the `foo` variable match would end. With this solution, the match will end at the first occurrence of a `b`